Converted all standard c types to types defined in stdint.h.
Performed comparisons of the hex file before and after the changes and the resulting hex was identical for both the rx and tx code.

There were two exceptions to this.  When the local fb inside 'rfmSetCarrierFrequency' and 'beacon_send' were converted to uint16_t from 'unsigned short' there was an unexplained difference in the resulting binary. Suspect some sort of type promotion by the compiler.

fb was left typed as 'unsigned short' in both cases until the cause of the difference is investigated.