Enable modeswitch support on 'metal case' deluxeTX.
The mode switch is connected to GND,SCL and SDA pins.

The GND is connected to the switch side pins and SCL and SDA to sides.
Both 2 and 3 position switches can be used.

Naturally other functions of the SCL/SDA pins cannot be used when switch
is connected to them.

Version -> 3.8.4.